requirements.txt
algoliasearch==2.6.2
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
atlassian-python-api==3.25.0
atomicwrites==1.4.0
attrs==21.2.0
Authlib==1.0.1
autopep8==1.5.7
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.9.3
bleach==4.1.0
blinker==1.5
bootstrap4==0.1.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
configparser==5.3.0
contentful==1.13.1
crispy-tailwind==0.5.0
cryptography==37.0.2
datafaker==0.7.6
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.6
django-appconf==1.0.5
django-bootstrap-datepicker-plus==4.0.0
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-filter==21.1
django-flatpickr==1.0.1
django-js-asset==1.2.2
django-model-utils==4.2.0
django-mptt==0.13.4
django-notification==1.3.3
django-notifications-hq==1.6.0
django-phonenumber-field==6.0.0
django-user-accounts==3.1.0
djangorestframework==3.13.1
entrypoints==0.3
et-xmlfile==1.1.0
Faker==2.0.2
fanstatic==1.2
fastapi==0.85.1
filelock==3.4.2
ghp-import==2.1.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.3
gunicorn==20.1.0
h11==0.13.0
harperdb==0.4.2
idna==3.2
importlib-metadata==4.11.4
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jira==3.2.0
js==1.0
jsonfield==3.1.0
jsonpickle==2.2.0
jsonschema==4.3.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.6.0
lxml==4.9.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
memory-profiler==0.60.0
mergedeep==1.3.4
mistune==0.8.4
mkdocs==1.3.0
mkdocs-material==8.2.16
mkdocs-material-extensions==1.0.3
multiaddr==0.0.9
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
netaddr==0.8.0
notebook==6.4.6
numpy==1.23.4
oauthlib==3.2.0
openpyxl==3.0.9
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.20.1
psutil==5.9.1
psycopg2==2.9.3
py==1.11.0
pyarrow==9.0.0
pycodestyle==2.7.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pydeck==0.8.0b1
Pygments==2.12.0
pymdown-extensions==9.4
Pympler==1.0.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.1.2
python-cfonts==1.5.2
python-dateutil==2.8.2
python-decouple==3.4
python-digitalocean==1.17.0
pytz==2021.1
# pywin32==303
# pywin32-ctypes==0.2.0
# pywinpty==1.1.6
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.4.1
rich-click==1.3.0
scipy==1.9.2
selenium==4.2.0
semantic-version==2.10.0
semver==2.13.0
Send2Trash==1.8.0
shutilwhich==1.1.0
simple-salesforce==1.11.6
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==1.3.8
sqlalchemy-cockroachdb==1.4.3
sqlparse==0.3.0
starlette==0.20.4
streamlit==1.12.2
swapper==1.3.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
traitlets==5.1.1
trio==0.21.0
trio-websocket==0.9.2
typer==0.4.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
Unipath==1.1
urllib3==1.26.6
uvicorn==0.18.3
validators==0.20.0
varint==1.0.2
virtualenv==20.13.0
vyper==0.3.7
waitress==2.0.0
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
wget==3.2
whitenoise==5.3.0
widgetsnbextension==3.5.2
wrapt==1.14.1
wsproto==1.1.0
you-get==0.4.1612
zeep==4.1.0
zipp==3.8.0
